requirements.txt
aiogram==3.4.1
SQLAlchemy==2.0.29
aiosqlite==0.19.0
aiohttp==3.9.5
python-dotenv==1.0.1